requirements.txt
acres==0.3.0
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.1.8
configobj==5.0.9
configparser==7.2.0
cryptography==44.0.3
etelemetry==0.3.1
filelock==3.18.0
fitz==0.0.1.dev2
Flask==3.1.0
Flask-Reuploaded==1.4.0
Flask-Uploads==0.2.1
Flask-WTF==1.2.2
google==3.0.0
google-auth==2.40.0
google-genai==1.13.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
looseversion==1.3.0
lxml==5.4.0
MarkupSafe==3.0.2
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pathlib==1.0.1
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
prov==2.0.1
puremagic==1.29
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydot==4.0.0
pyparsing==3.2.3
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyxnat==1.6.3
rdflib==6.3.2
requests==2.32.3
rsa==4.9.1
scipy==1.15.2
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
traits==7.0.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
websockets==15.0.1
Werkzeug==3.1.3
WTForms==3.2.1
